Add collision logic to map
Map now knows when an update that will result in a collision is
being attempted and will raise a `CollisionError` exception with
the type of the object subject to collision.

Objects have different types because not all collision pairs will
result in the same outcome. The game should know which object is
colliding with which object.