Snake can grow
Snake can now eat food and grow. This happens via the map
detecting a collision between snake and an object of type 'food'.
When the snake collides with a food, the game detects this and
grows the head of the snake at the place of the food. After this,
there is no need to explicitly move the snake on the map again
because it looks like it has already moved.